Cache the list of all projects
Cached in a global variable, so the whole list of projects doesn't
have to be re-fetched, given that it can take a long time with many
projects.

There is currently no cache expire except restart of achoo - perhaps
something for the future?

There are probably other things that could benefit from some caching,
but this is the most obvious one to me...